+++ pyoxidizer/src/projectmgmt.rs
@@ -314,7 +314,7 @@ pub fn init_rust_project(project_path: &Path) -> Result<()> {
     let env = crate::environment::resolve_environment()?;
     let pyembed_location = env.as_pyembed_location();
 
-    initialize_project(project_path, &pyembed_location, None, &[])?;
+    initialize_project(project_path, &pyembed_location, None, &[], "console")?;
     println!();
     println!(
         "A new Rust binary application has been created in {}",